backend/requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Base de datos
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Validación y utilidades
email-validator==2.1.0
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0

# Monitoring y logging
python-json-logger==2.0.7

# CORS y middleware
fastapi-limiter==0.1.5
redis==4.6.0

# Manejo de fechas
python-dateutil==2.8.2

# Generación de reportes
reportlab==4.0.9
openpyxl==3.1.2

# Manejo de imágenes
Pillow==10.2.0

# Background tasks
celery==5.3.6
